fix: ✏️ 총 지출 총합 값의 정수 범위 초과 케이스를 고려하여 타입 수정 (#133)
* fix: target_amount_dto 원시 타입으로 수정

* fix: spending_search_res daily_total_amount int -> long

* fix: daily_total_amount total_spending 계산 시, 반환 타입 long으로 수정

* fix: recent_target_amount_res 필드 원시 타입으로 수정

* style: spending_service 총 지출 금액 조회 메서드 인접하게 위치 수정

* fix: 총 지출 금액 dto amount 필드 int -> long

* style: target_amount_usecase 함수 호출 스니펫과 return 문 사이 공백 추가

* fix: 가장 최근 목표 금액의 amount가 0인 경우가 존재할 수 있으므로, 타입을 래퍼 타입으로 수정

* feat: total_spending_amount year_month 반환 메서드 추가

* fix: int 타입 long으로 변환 및 코드 정리

* fix: target amount update int -> long 타입 수정

* fix: querydsl total_spending_amount 기본 타입 expression 정의

* test: target amount controller unit test get_id() null 에러 핸들링

* test: max integer 범위 초과한 경우 diff_amount, total_spending 출력 확인

* fix: to_year_month_map 동시성 환경을 고려하여, concurrent_map 타입으로 반환

+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/ledger/dto/SpendingSearchRes.java
@@ -51,7 +51,7 @@ public class SpendingSearchRes {
             @Schema(description = "일")
             int day,
             @Schema(description = "일별 총 지출 금액")
-            int dailyTotalAmount,
+            long dailyTotalAmount,
             @Schema(description = "개별 지출 내역")
             List<Individual> individuals
     ) {
